coldoutbound/dataextraction/requirements.txt
flask==2.3.3
werkzeug==2.3.7
selenium==4.12.0
supabase==1.0.3
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.0
